requirements.txt
Flask
python-twitter
flask-cors